Allow newline-terminated classes for blocks
i.e. allow the user to forget the closing square bracket for classes.
This prevents something like
  ````some-id[some-class whoops no closing square bracket
    Intended code content (1)
  ````

  Accidental code content

  ````some-id[another-class]
    Intended code content [2]
  ````
from being interpreted as a single code block
with class "some-class whoops ... Intended code content [2",
at the cost of not being able to have classes with newlines in them.
If you really want a class with newlines in it, use CMD literals.