Improve placeholder storage (use X{n}X).
Let X denote U+E000, the first "Private Use Area" code point;
this is the "placeholder marker".

Previously, placeholder strings were of the form XX...X{n}X
where XX...X was a run of length one more than
the number of occurrences of X in the original markup.
Actually this is not a robust choice,
since inclusions (+ +) may inject a longer run.

Instead, it is better to use placeholder strings of the form X{n}X,
and use X0X as a placeholder for actual occurrences of X itself,
in both the CMD file and any inclusions.
This will work even when there is a longer run of Xs in an inclusion
than in the main CMD file being converted.