Initial commit for branch 'exception-handling'
Currently errors result in the built-in traceback.
When this corresponds to a nonexistent CMD filename
passed to cmd_file_to_html_file, this is alright,
but when the error is for an inclusion of a nonexistent file
or a broken regex replacement,
the traceback does not tell the user
where the error arose in terms of a CMD file.
In this branch I shall append an informative message
which tells the user which CMD file the error arose in
(and possibly the line number therein).
Since the built-in traceback should not be thrown away,
Python 3 exception chaining `raise ... from ... ` shall be used,
resulting in:

    Traceback (built-in traceback)

    The above exception was the direct cause of the following exception:

    Traceback (informative message)

See <https://stackoverflow.com/a/24752607>.