Fix attribute dictionary KeyError
This is the proper fix in the place of a280eb4d,
which was reverted in 1afcc18d.
The attribute dictionary returned by
get_definition_attribute_dictionary must be a deep copy,
otherwise the attribute dictionary as stored in ReferenceStorage
will be inadvertently altered by multiple references
unto the same link definition.