Re-implement inline semantics
The existing implementation is over-complicated and doesn't work.
In particular, the 33 delimiter case is processed before any other cases,
leading to unexpected behaviour.
For example,
    *1 **2*** A ***3** 4*
will produce
    <em>1 <strong>2<strong><em>A</em></strong>3</strong> 4</em>
because '*** A ***' is matched first,
whereas the intended output is
    <em>1 <strong>2</strong></em> A <em><strong>3</strong> 4</em>.
The new implementation does away with the nesting cases
and simply uses multiple passes until the markup stops changing,
with <CONTENT> only matched during each pass if it does not contain
the delimiting character.
Note that in the new implementation, delimiter matching is inner-greedy.
For example, ***blah*** produces
    <em><strong>blah</strong></em>
rather than
    <strong><em>blah</em></strong>.
Also, the opening delimiter must not be followed by whitespace or </,
and the closing delimiter must not be preceded by whitespace.